API Reference

Element.ButtonState — Type

Enum representing the state of a mouse button.

IsReleased: The button is currently released.

IsPressed: The button is currently pressed.

Element.MouseButton — Type

Enum representing the different mouse buttons.

LeftButton: The left mouse button.

RightButton: The right mouse button.

MiddleButton: The middle mouse button (scroll button).

Element.Container — Function

The Container is the most basic GUI component that can contain another component. It is the most basic building block of the GUI system.

Element.apply_layout — Method

apply_layout(component::AbstractView)

Apply layout to a GUI component and its children. This function calculates and applies the layout to components. The interpret_view function then uses the positions and sizes calculated by this function.

Element.draw_closed_lines — Method

draw_closed_lines(vertices::Vector{Point2f}, color_rgba::Vec4{<:AbstractFloat})

Draw closed lines using the provided vertices and color.

Element.draw_rectangle — Method

draw_rectangle(vertices::Vector{Point2f}, color_rgba::Vec4{<:AbstractFloat}, projection_matrix::Mat4{Float32})

Draw a rectangle using the provided vertices and color.

Element.generate_rectangle_vertices — Method

generate_rectangle_vertices(x, y, width, height)

Function to generate a rectangle with specified position and size in pixel coordinates.

This function creates a rectangle defined by its top-left corner (x, y), width, and height.

Element.get_orthographic_matrix — Method

get_orthographic_matrix(left::T, right::T, bottom::T, top::T, near::T, far::T)::Matrix{T} where {T<:Real}

Create an orthographic projection matrix.

Element.initialize_shaders — Method

Initialize the shader program (must be called after OpenGL context is created)

Element.interpret_view — Method

interpret_view(component::AbstractView, x::Float32, y::Float32, width::Float32, height::Float32, projection_matrix::Mat4{Float32})

Interpret the view of a GUI component. This function is responsible for interpreting the view of a GUI component based on its layout and properties.

Element.run — Method

run(ui_ref[]::AbstractView; title::String="Element", window_width_px::Integer=1920, window_height_px::Integer=1080)

Run the main loop for the GUI application. This function handles the rendering and event processing for the GUI.